refactor: split Game component
